Extract v1 discovery flow behind provider adapter seam
The version seam discovery has been missing: DiscoveryProviderAdapter
(start/stop/resume/cancel) with DiscoveryProviderAdapterFactory
routing on the run's connector-interface association. No association
routes to v1 — the deliberate inverse of the authority factory's
NULL rule, because discovery's legacy is the v1 interface — a v2
interface routes to the placeholder adapter that arrives for real
with the lifecycle implementation, and an unrecognized version
refuses rather than guessing.

V1DiscoveryProviderAdapter is the current flow moved verbatim out of
the version-blind DiscoveryServiceImpl: provider start, wait loop,
page download, duplicate handling and processing handoff, method
bodies unchanged. The service shrinks to orchestration and CRUD and
routes runDiscovery through the factory with a null-tolerant read,
so a run that cannot be loaded keeps taking the v1 flow's
established failure path. The factory's routing table is pinned by
unit tests; every existing discovery ITest passes unchanged.
